# Decoding performance optimization