--- a/.codeclimate.yml
+++ b/.codeclimate.yml
@@ -23,3 +23,4 @@ ratings:
   - "**.rb"
 exclude_paths:
 - test/
+- benchmark/