--- a/base62.js
+++ b/base62.js
@@ -1,4 +1,6 @@
 module.exports = (function (Base62) {
+    "use strict";
+
     var DEFAULT_CHARACTER_SET = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
     Base62.encode = function(integer){
@@ -11,7 +13,30 @@ module.exports = (function (Base62) {
         return s;
     };
 
-    Base62.decode = function(base62String){
+    var defaultCharsetDecode = function defaultCharsetDecode (base62String) {
+        var value = 0,
+            i = 0,
+            length = base62String.length,
+            charValue;
+
+        for (; i < length; i++) {
+            charValue = base62String.charCodeAt(i);
+
+            if (charValue < 58) {
+                charValue = charValue - 48;
+            } else if (charValue < 91) {
+                charValue = charValue - 29;
+            } else {
+                charValue = charValue - 87;
+            }
+
+            value += charValue * Math.pow(62, length - i - 1);
+        }
+
+        return value;
+    };
+
+    var customCharsetDecode = function customCharsetDecode (base62String) {
         var val = 0, base62Chars = base62String.split("").reverse();
         base62Chars.forEach(function(character, index){
             val += Base62.characterSet.indexOf(character) * Math.pow(62, index);
@@ -19,18 +44,26 @@ module.exports = (function (Base62) {
         return val;
     };
 
+    var decodeImplementation = null;
+
+    Base62.decode = function(base62String){
+        return decodeImplementation(base62String);
+    };
+
     Base62.setCharacterSet = function(chars) {
         var arrayOfChars = chars.split(""), uniqueCharacters = [];
 
-        if(arrayOfChars.length != 62) throw Error("You must supply 62 characters");
+        if(arrayOfChars.length !== 62) throw Error("You must supply 62 characters");
 
         arrayOfChars.forEach(function(char){
             if(!~uniqueCharacters.indexOf(char)) uniqueCharacters.push(char);
         });
 
-        if(uniqueCharacters.length != 62) throw Error("You must use unique characters.");
+        if(uniqueCharacters.length !== 62) throw Error("You must use unique characters.");
 
         Base62.characterSet = arrayOfChars;
+
+        decodeImplementation = chars === DEFAULT_CHARACTER_SET ? defaultCharsetDecode : customCharsetDecode;
     };
 
     Base62.setCharacterSet(DEFAULT_CHARACTER_SET);