The aim of this PR is to optimize the performance of the decode() method. The setCharacterSet method is now swapping the decode method implementation with an optimized one when the default charset is given. The improvement comes from the lack of array search (no `indexOf`) in this specific implementation. At first I intended to swap the implementation of Base62.decode directly but then I figured it would be an issue if anyone is storing the decode method in a variable ( `var b62decode = Base.decode;` ) , hence why the swap is done on the `decodeImplementation` variable.

The implementation of the decode method for custom charset has also been slightly optimized.

This mostly stems from my work on [ya-base62](https://github.com/kchapelier/ya-base62) and the benchmarking I did on it. That module has slightly different design goals (handling structured strings) and I figured out I might as well contribute to the most depended upon base62 module as it will probably be more beneficial to the community.

**Before modifications**

| Operations | Results | Elapsed time |
| ----------- | ------- | ------------- |
| decoding with default charset (1000000x) | correct | 2166.29 ms |
| encoding with default charset (1000000x) | correct | 58.42 ms |
| decoding with custom charset (1000000x) | correct | 2873.36 ms |
| encoding with custom charset (1000000x) | correct | 57.60 ms |

**After modifications**

| Operations | Results | Elapsed time |
| ----------- | ------- | ------------- |
| decoding with default charset (1000000x) | correct | 81.88 ms |
| encoding with default charset (1000000x) | correct | 58.63 ms |
| decoding with custom charset (1000000x) | correct | 1727.05 ms |
| encoding with custom charset (1000000x) | correct | 55.26 ms |

The benchmark code is included in the PR (it's excluded from `.codeclimate.yml`).

A few minor tests were added just to make sure unwanted changes are not introduced.