@andrew 2017-05-15
Wonderful, thanks for the contribution, I'll get a new release published shortly 🎉